chore: rename Post model
