+++ app/src/main/java/com/thinh/snaplet/data/datasource/remote/ApiService.kt
@@ -1,31 +1,31 @@
 package com.thinh.snaplet.data.datasource.remote
 
 import com.thinh.snaplet.data.model.BaseResponse
-import com.thinh.snaplet.data.model.CreatePostRequest
 import com.thinh.snaplet.data.model.EmailAvailabilityData
+import com.thinh.snaplet.data.model.FriendsCountData
 import com.thinh.snaplet.data.model.LoginRequest
 import com.thinh.snaplet.data.model.LoginResponse
-import com.thinh.snaplet.data.model.Post
-import com.thinh.snaplet.data.model.PostsFeedData
-import com.thinh.snaplet.data.model.UnreadPostsCountData
-import com.thinh.snaplet.data.model.MarkPostsSeenRequest
 import com.thinh.snaplet.data.model.RefreshTokenRequest
 import com.thinh.snaplet.data.model.RegisterRequest
-import com.thinh.snaplet.data.model.FriendsCountData
 import com.thinh.snaplet.data.model.Relationship
 import com.thinh.snaplet.data.model.RelationshipWithUserDto
-import com.thinh.snaplet.data.model.UpdateRelationshipRequest
 import com.thinh.snaplet.data.model.TokenResponse
+import com.thinh.snaplet.data.model.UpdateRelationshipRequest
 import com.thinh.snaplet.data.model.UsernameAvailabilityData
+import com.thinh.snaplet.data.model.media.ConfirmUploadData
+import com.thinh.snaplet.data.model.media.MediaConfirmUploadRequest
+import com.thinh.snaplet.data.model.media.RequestUploadRequest
+import com.thinh.snaplet.data.model.media.UploadRequestData
+import com.thinh.snaplet.data.model.post.CreatePostRequest
+import com.thinh.snaplet.data.model.post.MarkPostsSeenRequest
+import com.thinh.snaplet.data.model.post.Post
+import com.thinh.snaplet.data.model.post.PostsFeedData
+import com.thinh.snaplet.data.model.post.UnreadPostsCountData
 import com.thinh.snaplet.data.model.user.AvatarUploadRequest
 import com.thinh.snaplet.data.model.user.AvatarUploadRequestResponse
 import com.thinh.snaplet.data.model.user.ConfirmAvatarUploadRequest
 import com.thinh.snaplet.data.model.user.UpdateDisplayNameRequest
 import com.thinh.snaplet.data.model.user.UserProfile
-import com.thinh.snaplet.data.model.media.ConfirmUploadData
-import com.thinh.snaplet.data.model.media.MediaConfirmUploadRequest
-import com.thinh.snaplet.data.model.media.RequestUploadRequest
-import com.thinh.snaplet.data.model.media.UploadRequestData
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.DELETE
