+++ app/src/main/java/com/thinh/snaplet/data/repository/MediaRepository.kt
@@ -1,8 +1,8 @@
 package com.thinh.snaplet.data.repository
 
-import com.thinh.snaplet.data.model.Post
 import com.thinh.snaplet.data.model.media.ConfirmUploadData
 import com.thinh.snaplet.data.model.media.UploadRequestData
+import com.thinh.snaplet.data.model.post.Post
 import com.thinh.snaplet.utils.network.ApiResult
 
 interface MediaRepository {
