+++ app/src/main/java/com/thinh/snaplet/data/repository/post/PostRepositoryImpl.kt
@@ -1,17 +1,17 @@
 package com.thinh.snaplet.data.repository.post
 
 import com.thinh.snaplet.data.datasource.remote.ApiService
-import com.thinh.snaplet.data.model.MarkPostsSeenRequest
-import com.thinh.snaplet.data.model.Post
-import com.thinh.snaplet.data.model.PostsFeedData
-import com.thinh.snaplet.data.model.UnreadPostsCountData
+import com.thinh.snaplet.data.model.post.MarkPostsSeenRequest
 import com.thinh.snaplet.data.model.post.NewPostUpdate
+import com.thinh.snaplet.data.model.post.Post
+import com.thinh.snaplet.data.model.post.PostsFeedData
+import com.thinh.snaplet.data.model.post.UnreadPostsCountData
 import com.thinh.snaplet.platform.socket.SocketEvent
 import com.thinh.snaplet.platform.socket.SocketManager
 import com.thinh.snaplet.utils.Logger
 import com.thinh.snaplet.utils.network.ApiResult
-import com.thinh.snaplet.utils.network.safeApiCall
 import com.thinh.snaplet.utils.network.GsonHolder.gson
+import com.thinh.snaplet.utils.network.safeApiCall
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.mapNotNull
