+++ app/src/main/java/com/thinh/snaplet/domain/feed/FetchNewerFeedUseCase.kt
@@ -1,6 +1,6 @@
 package com.thinh.snaplet.domain.feed
 
-import com.thinh.snaplet.data.model.Post
+import com.thinh.snaplet.data.model.post.Post
 import com.thinh.snaplet.data.repository.post.PostRepository
 import com.thinh.snaplet.domain.feed.GetNewsfeedUseCase.Companion.FEED_PAGE_LIMIT
 import com.thinh.snaplet.domain.model.NewerFeedResult
