+++ app/src/main/java/com/thinh/snaplet/domain/model/UploadPostResult.kt
@@ -1,6 +1,6 @@
 package com.thinh.snaplet.domain.model
 
-import com.thinh.snaplet.data.model.Post
+import com.thinh.snaplet.data.model.post.Post
 
 sealed class UploadPostResult {
     data class Success(val post: Post) : UploadPostResult()
